Removed unnecessary Windows specific dependancy. (`corecrt.h`)

+++ vec.c
@@ -1,4 +1,3 @@
-#include <corecrt.h>
 #include <stdarg.h>
 #include <stdint.h>
 #include <stdio.h>
@@ -865,7 +864,7 @@ void vec_read_ascii_line_s(Vector *vec, FILE *file) {
 }
 
 // TODO: Remake this function.
-int vec_read_file(Vector *vec, char file_name[], size_t *bytes_written, bool minimize) {
+uint8_t vec_read_file(Vector *vec, char file_name[], size_t *bytes_written, bool minimize) {
     FILE *file;
     int err = 0;
     void *data;
